+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- SoftStaking nodes rewards
 
 ## [v2.28.2] 2021-09-22
 ### Changed